feat: refactor everything into components
