+++ frontend/components/Header.tsx
@@ -1,0 +1,9 @@
+export default function Header() {
+  return (
+    <div className="flex items-center justify-center sm:justify-between mb-6 sm:mb-8">
+      <h1 className="text-xl sm:text-2xl lg:text-3xl font-bold bg-gradient-to-r from-cyan-400 to-violet-400 bg-clip-text text-transparent hover:from-orange-400 hover:to-fuchsia-400 transition-all duration-500 cursor-pointer drop-shadow-[0_0_15px_rgba(34,211,238,0.6)]">
+        gitbeat 🎧
+      </h1>
+    </div>
+  );
+}
