+++ frontend/components/SongCard.tsx
@@ -1,0 +1,326 @@
+import { Play, Pause, ArrowUp } from "lucide-react";
+import { Song } from './types';
+
+interface SongCardProps {
+  song: Song;
+  index: number;
+  songs: Song[];
+  playingId: string | null;
+  isPlaying: boolean;
+  currentTime: number;
+  duration: number;
+  recentlyUpvoted: Set<string>;
+  rankingChanges: Set<string>;
+  rankingDirections: {[key: string]: 'up' | 'down'};
+  onTogglePlay: (id: string) => void;
+  onUpvote: (id: string) => void;
+  onSeek: (time: number) => void;
+  formatTime: (time: number) => string;
+}
+
+export default function SongCard({
+  song,
+  index,
+  songs,
+  playingId,
+  isPlaying,
+  currentTime,
+  duration,
+  recentlyUpvoted,
+  rankingChanges,
+  rankingDirections,
+  onTogglePlay,
+  onUpvote,
+  onSeek,
+  formatTime
+}: SongCardProps) {
+  const getRankStyle = (index: number) => {
+    if (index === 0) return 'bg-yellow-500 text-black shadow-lg shadow-yellow-500/50';
+    if (index === 1) return 'bg-gray-400 text-black shadow-lg shadow-gray-400/50';
+    if (index === 2) return 'bg-amber-600 text-black shadow-lg shadow-amber-600/50';
+    return 'text-white';
+  };
+
+  const getPopularityPercentage = () => {
+    const maxUpvotes = Math.max(...songs.map(s => s.upvote_count), 1);
+    return Math.round((song.upvote_count / maxUpvotes) * 100);
+  };
+
+  return (
+    <div
+      className={`p-3 sm:p-4 rounded-lg transition-all duration-500 hover:shadow-xl transform ${
+        rankingChanges.has(song.id) 
+          ? 'animate-pulse scale-105 ring-2 ring-emerald-400/50'
+          : ''
+      } bg-[#1C2530] hover:bg-slate-800 border-slate-600 hover:border-emerald-500/30 shadow-[0_0_25px_rgba(16,185,129,0.1)] hover:shadow-emerald-500/20 border hover:scale-[1.02]`}
+      style={{
+        transitionProperty: 'all, transform, box-shadow',
+        transitionDuration: '800ms',
+        transitionTimingFunction: 'cubic-bezier(0.4, 0, 0.2, 1)',
+        animationDelay: `${index * 100}ms`,
+      }}
+    >
+      {/* Mobile Layout */}
+      <div className="block sm:hidden">
+        {/* Top row: Rank, Title, Actions */}
+        <div className="flex items-center justify-between mb-3">
+          <div className="flex items-center gap-3 flex-1 min-w-0">
+            {/* Rank */}
+            <div className="relative flex-shrink-0">
+              <div className={`w-6 h-6 rounded-lg flex items-center justify-center text-xs font-bold transition-all duration-500 ${
+                rankingChanges.has(song.id) ? 'animate-bounce scale-125' : ''
+              } ${getRankStyle(index)}`}>
+                {index + 1}
+              </div>
+              {rankingChanges.has(song.id) && rankingDirections[song.id] && (
+                <div className={`absolute -top-1 -right-1 w-3 h-3 rounded-full flex items-center justify-center text-xs animate-ping ${
+                  rankingDirections[song.id] === 'up' 
+                    ? 'bg-green-500 text-white' 
+                    : 'bg-red-500 text-white'
+                }`}>
+                  {rankingDirections[song.id] === 'up' ? '↑' : '↓'}
+                </div>
+              )}
+            </div>
+            
+            {/* Song info */}
+            <div className="flex-1 min-w-0">
+              <div className="font-semibold text-white hover:text-emerald-400 transition-colors drop-shadow-[0_0_5px_rgba(16,185,129,0.3)] text-sm truncate">
+                {song.title || `${song.repository.name} Beat`}
+              </div>
+              <a
+                href={song.repository.url}
+                target="_blank"
+                rel="noopener noreferrer"
+                className="text-xs text-slate-400 hover:text-emerald-400 transition-colors hover:drop-shadow-[0_0_5px_rgba(16,185,129,0.5)] block truncate"
+              >
+                {song.repository.url.replace('https://github.com/', '')}
+              </a>
+              {song.lyrics_url && (
+                <a
+                  href={song.lyrics_url}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                  className="block text-xs text-violet-400 hover:text-violet-300 transition-colors"
+                >
+                  View Lyrics
+                </a>
+              )}
+            </div>
+          </div>
+          
+          {/* Action buttons */}
+          <div className="flex items-center gap-2 flex-shrink-0">
+            <button
+              onClick={() => onTogglePlay(song.id)}
+              className="hover:cursor-pointer w-8 h-8 text-white rounded-md hover:bg-emerald-400 hover:scale-105 transition-all duration-300 flex items-center justify-center"
+            >
+              {playingId === song.id && isPlaying ? (
+                <Pause size={14} />
+              ) : (
+                <Play size={14} />
+              )}
+            </button>
+            <button
+              onClick={() => onUpvote(song.id)}
+              className={`hover:cursor-pointer w-8 h-8 text-black rounded-md transition-all duration-300 flex items-center justify-center shadow-[0_0_15px_rgba(139,92,246,0.3)] ${
+                recentlyUpvoted.has(song.id)
+                  ? 'bg-violet-400 scale-110 shadow-lg shadow-violet-500/50 animate-pulse'
+                  : 'bg-emerald-300 hover:bg-violet-400 hover:scale-105 hover:shadow-lg hover:shadow-violet-500/50'
+              }`}
+            >
+              <ArrowUp size={14} className={recentlyUpvoted.has(song.id) ? 'animate-bounce' : ''} />
+            </button>
+          </div>
+        </div>
+        
+        {/* Bottom row: Stats */}
+        <div className="flex items-center justify-between">
+          <div className="flex items-center gap-6">
+            <div className="text-center">
+              <div className={`text-sm font-bold text-white hover:text-violet-300 transition-all duration-300 drop-shadow-[0_0_8px_rgba(139,92,246,0.6)] ${
+                recentlyUpvoted.has(song.id) ? 'scale-125 text-violet-300' : ''
+              }`}>
+                {song.upvote_count}
+              </div>
+              <div className="text-xs text-slate-400">upvotes</div>
+            </div>
+            <div className="text-center">
+              <div className="text-sm font-bold text-emerald-400 hover:text-emerald-300 transition-colors drop-shadow-[0_0_8px_rgba(16,185,129,0.6)]">
+                {song.audio_url ? '🎵' : '⏳'}
+              </div>
+              <div className="text-xs text-slate-400">audio</div>
+            </div>
+          </div>
+          
+          <div className="text-right">
+            <div className="text-sm font-bold text-emerald-300">
+              {getPopularityPercentage()}%
+            </div>
+            <div className="text-xs text-slate-400">popularity</div>
+          </div>
+        </div>
+      </div>
+
+      {/* Desktop Layout */}
+      <div className="hidden sm:flex items-center">
+        {/* Rank */}
+        <div className="relative flex-shrink-0">
+          <div className={`w-8 h-8 rounded-lg flex items-center justify-center text-sm font-bold transition-all duration-500 ${
+            rankingChanges.has(song.id) ? 'animate-bounce scale-125' : ''
+          } ${getRankStyle(index)}`}>
+            {index + 1}
+          </div>
+          {rankingChanges.has(song.id) && rankingDirections[song.id] && (
+            <div className={`absolute -top-2 -right-2 w-4 h-4 rounded-full flex items-center justify-center text-xs animate-ping ${
+              rankingDirections[song.id] === 'up' 
+                ? 'bg-green-500 text-white' 
+                : 'bg-red-500 text-white'
+            }`}>
+              {rankingDirections[song.id] === 'up' ? '↑' : '↓'}
+            </div>
+          )}
+        </div>
+
+        {/* Song info */}
+        <div className="flex items-center gap-3 ml-4 min-w-[200px]">
+          <div>
+            <div className="font-semibold text-white hover:text-emerald-400 transition-colors drop-shadow-[0_0_5px_rgba(16,185,129,0.3)]">
+              {song.title || `${song.repository.name} Beat`}
+            </div>
+            <a
+              href={song.repository.url}
+              target="_blank"
+              rel="noopener noreferrer"
+              className="text-xs text-slate-400 hover:text-emerald-400 transition-colors hover:drop-shadow-[0_0_5px_rgba(16,185,129,0.5)]"
+            >
+              {song.repository.url.replace('https://github.com/', '')}
+            </a>
+            {song.lyrics_url && (
+              <a
+                href={song.lyrics_url}
+                target="_blank"
+                rel="noopener noreferrer"
+                className="block text-xs text-violet-400 hover:text-violet-300 transition-colors"
+              >
+                View Lyrics
+              </a>
+            )}
+          </div>
+        </div>
+
+        {/* Stats */}
+        <div className="flex items-center gap-4 lg:gap-8 ml-4 lg:ml-8">
+          <div className="text-center">
+            <div className={`text-sm lg:text-lg font-bold text-white hover:text-violet-300 transition-all duration-300 drop-shadow-[0_0_8px_rgba(139,92,246,0.6)] ${
+              recentlyUpvoted.has(song.id) ? 'scale-125 text-violet-300' : ''
+            }`}>
+              {song.upvote_count}
+            </div>
+            <div className="text-xs text-slate-400">upvotes</div>
+          </div>
+          <div className="text-center">
+            <div className="text-sm lg:text-lg font-bold text-emerald-400 hover:text-emerald-300 transition-colors drop-shadow-[0_0_8px_rgba(16,185,129,0.6)]">
+              {song.audio_url ? '🎵' : '⏳'}
+            </div>
+            <div className="text-xs text-slate-400">audio</div>
+          </div>
+        </div>
+
+        {/* Progress bar */}
+        <div className="flex-1 mx-4 lg:mx-8 min-w-[100px]">
+          <div className="flex justify-between text-xs text-slate-400 mb-1">
+            <span>popularity</span>
+            <span>{getPopularityPercentage()}%</span>
+          </div>
+          <div className="w-full bg-slate-700 rounded-lg h-2 overflow-hidden shadow-inner">
+            <div 
+              className="bg-emerald-300 h-2 rounded-lg hover:from-fuchsia-500 hover:to-violet-500 transition-all duration-1000 shadow-[0_0_10px_rgba(16,185,129,0.5)]" 
+              style={{ width: `${getPopularityPercentage()}%` }}
+            ></div>
+          </div>
+        </div>
+
+        {/* Action buttons */}
+        <div className="flex items-center gap-2">
+          <button
+            onClick={() => onTogglePlay(song.id)}
+            className="hover:cursor-pointer w-10 h-10 text-white rounded-md hover:bg-emerald-400 hover:scale-105 transition-all duration-300 flex items-center justify-center"
+          >
+            {playingId === song.id && isPlaying ? (
+              <Pause size={16} />
+            ) : (
+              <Play size={16} />
+            )}
+          </button>
+          <button
+            onClick={() => onUpvote(song.id)}
+            className={`hover:cursor-pointer w-10 h-10 text-black rounded-md transition-all duration-300 flex items-center justify-center shadow-[0_0_15px_rgba(139,92,246,0.3)] ${
+              recentlyUpvoted.has(song.id)
+                ? 'bg-violet-400 scale-110 shadow-lg shadow-violet-500/50 animate-pulse'
+                : 'bg-emerald-300 hover:bg-violet-400 hover:scale-105 hover:shadow-lg hover:shadow-violet-500/50'
+            }`}
+          >
+            <ArrowUp size={16} className={recentlyUpvoted.has(song.id) ? 'animate-bounce' : ''} />
+          </button>
+        </div>
+      </div>
+
+      {/* Enhanced Playing indicator with controls */}
+      {playingId === song.id && (
+        <div className="mt-4 pt-4 border-t border-slate-700">
+          <div className="flex items-center justify-between mb-2">
+            <div className="text-xs text-emerald-400 drop-shadow-[0_0_10px_rgba(16,185,129,0.8)]">
+              {isPlaying ? '♪ playing beat...' : 'paused'}
+            </div>
+            <div className="text-xs text-slate-400">
+              {formatTime(currentTime)} / {formatTime(duration)}
+            </div>
+          </div>
+          
+          {/* Seekable progress bar */}
+          <div 
+            className="w-full bg-slate-700 rounded-lg h-2 overflow-hidden shadow-inner cursor-pointer relative group"
+            onClick={(e) => {
+              const rect = e.currentTarget.getBoundingClientRect();
+              const clickX = e.clientX - rect.left;
+              const width = rect.width;
+              const newTime = (clickX / width) * duration;
+              onSeek(newTime);
+            }}
+          >
+            {/* Progress bar */}
+            <div 
+              className="bg-gradient-to-r bg-emerald-300 h-2 rounded-lg transition-all duration-100 shadow-[0_0_15px_rgba(16,185,129,0.7)]"
+              style={{ width: `${duration > 0 ? (currentTime / duration) * 100 : 0}%` }}
+            ></div>
+            
+            {/* Hover indicator */}
+            <div 
+              className="absolute top-0 h-2 bg-white/20 opacity-0 group-hover:opacity-100 transition-opacity duration-200 pointer-events-none"
+              style={{ 
+                left: '0px',
+                width: '2px',
+                transform: 'translateX(var(--mouse-x, 0px))'
+              }}
+            ></div>
+          </div>
+          
+          {/* Playback controls */}
+          <div className="flex items-center justify-center mt-3 gap-2">
+            <button
+              onClick={() => onTogglePlay(song.id)}
+              className="w-8 h-8 bg-emerald-400 text-black rounded-full hover:bg-emerald-300 transition-all duration-200 flex items-center justify-center hover:scale-110 hover:cursor-pointer"
+            >
+              {isPlaying ? <Pause size={12} /> : <Play size={12} />}
+            </button>
+            
+            <div className="text-xs text-slate-400 ml-2">
+              Click on the progress bar to seek
+            </div>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+}
