+++ frontend/components/utils/github.ts
@@ -1,0 +1,8 @@
+export const isValidGitHubUrl = (url: string): boolean => {
+  try {
+    const urlObj = new URL(url);
+    return urlObj.hostname === 'github.com' && urlObj.pathname.split('/').length >= 3;
+  } catch {
+    return false;
+  }
+};
